add attempted fix to the issue
adds an attempted fix for this issue, where
the rest.currentuser would be set in the socket ready event.

this won't work because the socketcurrentuser is different from the
rest one iirc. in addition, it bodges on another method to the rest client
that seems like a poor quality solution